requirements.txt
customtkinter
psutil
Pillow
win10toast
matplotlib  # optional, for graphing
pystray
Pillow
pyinstaller